requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
Flask-Mailman==1.1.1
Flask-WTF==1.2.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
sqlmodel==0.0.31
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1
zipp==3.23.0
